reorganize structure of services
